Fix sleep timer bug
This will fix the current bug for the sleep timer. The sleep timer currently works find but if you enable it for 1 minute and then change it to 1 hour it will remember the old time., this is fixed.
Also the new way will handle the realtime change for the slider.
Tested.