UI rework / backend
This commit is a huge one. Too bad.
Redesign, adaptive ui for ios, blurred appbars. New player design. Removed landscape support (i don't know how to make it work).
Removed a bunch of customization and optimize the app by removing animated covers from the radio list.
This design should work for 99% of the devices and it's much better.
Added a new timer for the current song.